Вносит правки, выделяет addComment в отдельный модуль
