+++ js/comments-viewer.js
@@ -1,11 +1,12 @@
+import {addComment} from './comments-renderer.js';
+
 const socialCommentsElement = document.querySelector('.social__comments');
-const commentTemplateElement = socialCommentsElement.querySelector('.social__comment');
 const bigPictureElement = document.querySelector('.big-picture');
 const loadCommentElement = bigPictureElement.querySelector('.social__comment-count');
 const currentCommentsCountElement = bigPictureElement.querySelector('.social__comment-shown-count');
 const totalCommentsCountElement = bigPictureElement.querySelector('.social__comment-total-count');
 const loadCommentsButtonElement = bigPictureElement.querySelector('.social__comments-loader');
-const loadingStep = 5;
+const LOADING_STEP = 5;
 let commentsArray = [];
 let totalComments;
 let currentCommentsLoaded = 0;
@@ -18,7 +19,7 @@ function fillCommentsList(comments) {
     currentCommentsLoaded = 0;
     totalComments = comments.length;
 
-    const commentsToLoad = Math.min(loadingStep, totalComments);
+    const commentsToLoad = Math.min(LOADING_STEP, totalComments);
     for (let i = 0; i < commentsToLoad; i++) {
       addCommentAndUpdateCounter(commentsArray[i]);
     }
@@ -31,7 +32,7 @@ function fillCommentsList(comments) {
 }
 
 function loadMoreComments() {
-  const commentsToLoad = Math.min(loadingStep, totalComments - currentCommentsLoaded);
+  const commentsToLoad = Math.min(LOADING_STEP, totalComments - currentCommentsLoaded);
 
   for (let i = 0; i < commentsToLoad; i++) {
     addCommentAndUpdateCounter(commentsArray[currentCommentsLoaded]);
@@ -39,18 +40,9 @@ function loadMoreComments() {
   loadCommentsButtonElement.classList.toggle('hidden', currentCommentsLoaded >= totalComments);
 }
 
-function addComment({avatar, name, commentsList}) {
-  const newCommentElement = commentTemplateElement.cloneNode(true);
-  const commentPictureElement = newCommentElement.querySelector('.social__picture');
-
-  commentPictureElement.setAttribute('src', avatar);
-  commentPictureElement.setAttribute('alt', name);
-  newCommentElement.querySelector('.social__text').textContent = commentsList;
-  socialCommentsElement.appendChild(newCommentElement);
-}
-
 function addCommentAndUpdateCounter(comment) {
-  addComment(comment);
+  addComment(comment, socialCommentsElement);
+
   currentCommentsLoaded++;
   currentCommentsCountElement.textContent = currentCommentsLoaded;
   totalCommentsCountElement.textContent = totalComments;
